+++ foursight_core/app_utils.py
@@ -14,6 +14,7 @@ import logging
 from itertools import chain
 from dateutil import tz
 from dcicutils import ff_utils
+from dcicutils.cloudformation_utils import AbstractOrchestrationManager
 from dcicutils.lang_utils import disjoined_list
 from dcicutils.secrets_utils import (
     apply_identity,
@@ -65,7 +66,7 @@ class AppUtilsCore:
     TRIM_ERR_OUTPUT = 'Output too large to provide on main page - see check result directly'
     LAMBDA_MAX_BODY_SIZE = 5500000  # 6Mb is the "real" threshold
 
-    # dmichaels/2022-07-20: New to apply the IDENTITY values globally to os.environ (C4-826)
+    # dmichaels/2022-07-20/C4-826: New to apply the IDENTITY values globally to os.environ
     # from the global application configuraiton (GAC). Will count on getting these values:
     #
     # Foursight Name     GAC Name
@@ -95,6 +96,13 @@ class AppUtilsCore:
             raise Exception("Foursight {GLOBAL_APPLICATION_CONFIGURATION} environment variable not set!")
         logger.info(f"Foursight {GLOBAL_APPLICATION_CONFIGURATION} environment variable value is: {identity_name}")
 
+        # Make sure the STACK_NAME (environment variable) is set (via Foursight CloudFormation template);
+        # from this we are able to find the CHECK_RUNNER lambda function name.
+        stack_name = os.environ.get("STACK_NAME")
+        if not stack_name:
+            raise Exception("Foursight STACK_NAME environment variable not set!")
+        logger.info(f"Foursight STACK_NAME environment variable value is: {stack_name}")
+
         # Apply the GAC secrets values globally to os.environ.
         apply_identity(identity_kind=GLOBAL_APPLICATION_CONFIGURATION, rename_keys=IDENTITY_KEY_MAP)
 
@@ -107,7 +115,20 @@ class AppUtilsCore:
             rds_secrets = SecretsTable(rds_secrets_name)
             os.environ["RDS_NAME"] = rds_secrets.get("dbInstanceIdentifier")
 
-        # dmichaels/2022-07-20: Set to proxy for local testing (e.g. http://localhost:9200).
+        # Get the CHECK_RUNNER lambda function name, using the stack_name as a prefix; for example,
+        # this name looks like: c4-foursight-cgap-supertest-stack-CheckRunner-8pYO8J9Tzb2P
+        # where c4-foursight-cgap-supertest-stack is the stack_name.
+        check_runner_lambda_function_name_pattern = stack_name + "-CheckRunner-.*"
+        check_runner_lambda_function_names = (
+            AbstractOrchestrationManager.find_lambda_function_names(check_runner_lambda_function_name_pattern))
+        if not check_runner_lambda_function_names:
+            raise Exception("Foursight CheckRunner lambda not found: {check_runner_lambda_function_name_pattern}")
+        elif len(check_runner_lambda_function_names) != 1:
+            raise Exception("Unique Foursight CheckRunner lambda not found: {check_runner_lambda_function_name_pattern}")
+        os.environ["CHECK_RUNNER"] = check_runner_lambda_function_names[0]
+        logger.info(f"Foursight CHECK_RUNNER environment variable value is: {os.environ['CHECK_RUNNER']}")
+
+        # Set ES_HOST to proxy for local testing (e.g. http://localhost:9200) via ES_HOST_LOCAL environment variable.
         es_host_local = os.environ.get("ES_HOST_LOCAL")
         if es_host_local:
             os.environ["ES_HOST"] = es_host_local